refactor backround job
